Separate keeva and square-otto from soomla libs

+++ Soomla/Assets/Soomla/Scripts/Core/Config/SoomlaEditorScript.cs
@@ -260,6 +260,9 @@ namespace Soomla
 		{
 			List<string> folders = new List<string>();
 			foreach (string file in filePaths) {
+				if (file.EndsWith ("Common")) {
+					continue;
+				}
 				FileUtil.DeleteFileOrDirectory(file);
 				string folderPath = Path.GetDirectoryName(file);
 				do {
